requirements.txt
alembic==1.7.7
APScheduler==3.6.3
attrs==19.3.0
boto3==1.23.10
botocore==1.26.10
certifi==2019.11.28
charset-normalizer==2.0.12
Click==7.0
cloudpickle==2.1.0
colorhash==1.0.2
configparser==4.0.2
cycler==0.10.0
databricks-cli==0.16.6
docker==5.0.3
entrypoints==0.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
gitdb==4.0.9
GitPython==3.1.18
gunicorn==20.1.0
idna==3.3
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.0
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
Mako==1.1.6
MarkupSafe==1.1.1
matplotlib==3.1.2
mlflow==1.23.1
more-itertools==8.1.0
numpy==1.18.1
oauthlib==3.2.0
packaging==21.3
pandas==0.25.3
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
protobuf==3.19.4
psutil==5.6.7
PyJWT==2.4.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
querystring-parser==1.2.4
regexp==0.1
requests==2.27.1
s3transfer==0.5.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sklearn-pandas==1.8.0
smmap==5.0.0
SQLAlchemy==1.3.13
sqlparse==0.4.2
tabulate==0.8.9
typing_extensions==4.1.1
tzlocal==2.0.0
urllib3==1.26.9
websocket-client==1.3.1
Werkzeug==0.16.1
wincertstore==0.2
xgboost==0.90
zipp==3.6.0
